mycar/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
asgiref==3.2.10
atomicwrites==1.4.1
attrs==23.1.0
certifi==2023.11.17
chardet==3.0.4
colorama==0.4.6
Django==3.1
django-background-tasks==1.2.5
django-compat==1.0.15
django-debug-toolbar==2.2
Faker==12.0.1
frozenlist==1.4.0
funcy==2.0
idna==2.8
mixer==7.1.2
more-itertools==10.1.0
multidict==6.0.4
packaging==23.2
Pillow==9.4.0
pluggy==0.13.1
psycopg2==2.9.9
py==1.11.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.22.0
six==1.14.0
sorl-thumbnail==12.6.3
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.25.11
wcwidth==0.2.12
yarl==1.9.3
